g7_workspace/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
blas=1.0=mkl
bzip2=1.0.6=h470a237_2
ca-certificates=2018.03.07=0
certifi=2018.8.13=py36_0
cffi=1.11.5=py36h9745a5d_0
cudatoolkit=9.0=h13b8566_0
ffmpeg=4.0.2=ha6a6e2b_0
freetype=2.8.1=hfa320df_1
gmp=6.1.2=hfc679d8_0
gnutls=3.5.19=h2a4e5f8_1
intel-openmp=2018.0.3=0
jpeg=9b=h024ee3a_2
libedit=3.1.20170329=h6b74fdf_2
libffi=3.2.1=hd88cf55_4
libgcc-ng=8.2.0=hdf63c60_1
libgfortran-ng=7.2.0=hdf63c60_3
libiconv=1.15=h470a237_2
libpng=1.6.34=hb9fc6fc_0
libstdcxx-ng=8.2.0=hdf63c60_1
libtiff=4.0.9=he85c1e1_1
mkl=2018.0.3=1
mkl_fft=1.0.4=py36h4414c95_1
mkl_random=1.0.1=py36h4414c95_1
ncurses=6.1=hf484d3e_0
nettle=3.3=0
ninja=1.8.2=py36h6bb024c_1
numpy=1.15.0=py36h1b885b7_0
numpy-base=1.15.0=py36h3dfced4_0
olefile=0.45.1=py36_0
openh264=1.7.0=0
openssl=1.0.2p=h14c3975_0
pandas=0.23.4=py36h04863e7_0
pillow=5.2.0=py36_0
pip=10.0.1=py36_0
pycparser=2.18=py36_1
python=3.6.6=hc3d631a_0
python-dateutil=2.7.3=py36_0
pytorch=0.4.1=py36_cuda9.0.176_cudnn7.1.2_1
pytz=2018.5=py36_0
readline=7.0=ha6073c6_4
setuptools=40.0.0=py36_0
six=1.11.0=py36_1
sqlite=3.24.0=h84994c4_0
tk=8.6.7=hc745277_3
torchvision=0.2.1=py36_1
wheel=0.31.1=py36_0
x264=1!152.20180717=h470a237_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=ha838bed_2
